`numRows()` is _not_ testable, mark it for future
as per the PDO documentation, `rowCount()` is intended to return the
number of rows affected by a manipulative query, and certain drivers
support returning a number of selected rows during a `SELECT`. however,
we are testing using sqlite in-memory, and sqlite itself does not
support `rowCount()` for select, so skip code coverage for it.