work-in-progress for Statement Containers
a StatementContainer is a class which holds the object or resource
pertaining to a query statement. it is agnostic of type, but will
identify the contents. support for exotic freeing of statements is
supported (similar to how mysql_free_result() works, provide a callable
to the driver class to handle the implementation). PdoDriver got the
first conversion, DoctrineDbal is getting the next implementation.

this fails tests at the moment and is being checkpointed for repo
carriage.