:RAGEFACE: please use bits if you use bitwise ops
firstly, if you use the integer values hardcoded in your code, i'm not
sorry: you deserve to have stuff break. but this innocuous change is
brought about by the fact that DB_FETCHMODE_OBJECT = 3, ergo this is
synonymous with DB_FETCHMODE_ORDERED | DB_FETCHMODE_ASSOC.

really, this isn't worthy of being bitwise (or maybe it is, a flipped
object, perhaps?). but really this is frought with pain.